rabbitMq + duende

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -2,6 +2,10 @@ using AuctionService.Data;
 using AuctionService.DTOs;
 using AuctionService.Entities;
 using AutoMapper;
+using AutoMapper.QueryableExtensions;
+using Contracts;
+using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -13,21 +17,26 @@ namespace AuctionService.Controllers
   {
     private readonly AuctionDbContext _context;
     private readonly IMapper _mapper;
-    public AuctionsController(AuctionDbContext context, IMapper mapper)
+    private readonly IPublishEndpoint _publishEndpoint;
+
+    public AuctionsController(AuctionDbContext context, IMapper mapper, IPublishEndpoint publishEndpoint)
     {
       _mapper = mapper;
+      _publishEndpoint = publishEndpoint;
       _context = context;
     }
 
     [HttpGet]
-    public async Task<ActionResult<List<AuctionDto>>> GetAllAuctions()
+    public async Task<ActionResult<List<AuctionDto>>> GetAllAuctions(string date)
     {
-      var auctions = await _context.Auctions
-        .Include(x => x.Item)
-        .OrderBy(x => x.Item.Make)
-        .ToListAsync();
+      var query = _context.Auctions.OrderBy(x => x.Item.Make).AsQueryable();
 
-      return _mapper.Map<List<AuctionDto>>(auctions);
+      if (!string.IsNullOrWhiteSpace(date))
+      {
+        query = query.Where(x => x.UpdatedAt.CompareTo(DateTime.Parse(date).ToUniversalTime()) > 0);
+      }
+
+      return await query.ProjectTo<AuctionDto>(_mapper.ConfigurationProvider).ToListAsync();
     }
 
     [HttpGet("{id}")]
@@ -42,20 +51,29 @@ namespace AuctionService.Controllers
       return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
       var auction = _mapper.Map<Auction>(auctionDto);
-      //TODO: add current use as seller
-      auction.Seller = "test";
+
+      auction.Seller = User.Identity.Name;
+
       _context.Auctions.Add(auction);
 
+      var newAuction = _mapper.Map<AuctionDto>(auction);
+
+      await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
+
       var result = await _context.SaveChangesAsync() > 0;
+
       if (!result) return BadRequest("Could not save changes to the DB");
-      return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, _mapper.Map<AuctionDto>(auction));
+
+      return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
 
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, AuctionDto updateAuctionDto)
     {
@@ -64,7 +82,7 @@ namespace AuctionService.Controllers
 
       if (auction == null) return NotFound();
 
-      //TODO: Check seller == username
+      if (auction.Seller != User.Identity.Name) return Forbid();
 
       auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
       auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -72,12 +90,15 @@ namespace AuctionService.Controllers
       auction.Item.Mileage = updateAuctionDto.Mileage ?? auction.Item.Mileage;
       auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
 
+      await _publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auction));
+
       var result = await _context.SaveChangesAsync() > 0;
 
       if (result) return Ok();
       return BadRequest("Problem saving changes");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -85,9 +106,11 @@ namespace AuctionService.Controllers
 
       if (auction == null) return NotFound();
 
-      //TODO: Check seller
+      if (auction.Seller != User.Identity.Name) return Forbid();
 
       _context.Auctions.Remove(auction);
+
+      await _publishEndpoint.Publish(_mapper.Map<AuctionDeleted>(auction));
 
       var result = await _context.SaveChangesAsync() > 0;
